requirements.txt
# AI UI自动化测试平台依赖

# 基础依赖
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# 数据库依赖
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# 测试工具
playwright>=1.40.0

# 计算机视觉
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
pillow>=10.0.0

# 其他工具
requests>=2.31.0
python-dotenv>=1.0.0
pyautogui>=0.9.54

# 开发工具
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0

# 认证依赖
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# 监控依赖
psutil>=5.9.0
